Make path-match? more useful, add factory id to linked paths
When recursing into linked factories and adding these to the
`:facai.result/linked` set, store them with a path that includes the factory
id (symbol), in addition to the map-entry key by which they are referenced. This
allows us to match either on key, or on associated value/type.

To make this work update `path-match?`, it now checks the type of the final
entry in the selector, if it's a keyword then the trailing factory id in the
path is ignored, so we retain the old behavior, as well supporting matching by
factory id.

Handle some other edge cases and syntax features: sets for alternation, more
correct handling of `:*`.