Make a new key :facai.build/factory availabe on the context
This key is added before invoking {before,after}-build-factory hooks, as a way
to access the current factory that is being built. It's usually the same as
facai.build/input, but will be unwrapped in case of a deferred.

Also allow vars as thunks (which get dereffed), the idea being that you can
call (defer #'my-factory {... opts ...}). Not sure yet if this will stick.